Fix aside resize position bug
When the document loads in narrow mode and is resized to wide mode the
asides are all placed at the top of the document. This is because there
is no listener to update their calculated position in the document on a
window resize event. Hence the on load runs once and sets the position
to 0. Add a listener to compute this the correct value on window resize.